+++ scripts/test/sample/SysConfig/Migration/ZZZAutoOTRS5.pm
@@ -845,7 +845,32 @@ $Self->{'Frontend::ToolBarModule'}->{'14-CICSearchCustomerUser'} =  {
   'Priority' => '1990040'
 };
 $Self->{'Ticket::Frontend::OverviewSmall'}->{'ColumnHeader'} =  'LastCustomerSubject';
-##+;
+$Self->{'PostMaster::PreFilterModule'}->{'1-Match'} =  {
+  'Match' => {
+    'From' => 'noreply@'
+  },
+  'Module' => 'Kernel::System::PostMaster::Filter::Match',
+  'Set' => {
+    'X-OTRS-ArticleType' => 'email-internal',
+    'X-OTRS-FollowUp-ArticleType' => 'email-external',
+    'X-OTRS-Ignore' => 'yes'
+  },
+};
+$Self->{'PostMaster::PreCreateFilterModule'}->{'000-FollowUpArticleTypeCheck'} =  {
+  'ArticleType' => 'email-internal',
+  'Module' => 'Kernel::System::PostMaster::Filter::FollowUpArticleTypeCheck',
+  'SenderType' => 'customer',
+  'X-OTRS-ArticleType' => 'email-internal',
+  'X-OTRS-FollowUp-ArticleType' => 'email-external'
+};
+$Self->{'PostMaster::CheckFollowUpModule'}->{'0100-Subject'} =  {
+  'ArticleType' => 'email-external',
+  'Module' => 'Kernel::System::PostMaster::FollowUpCheck::Subject',
+  'SenderType' => 'customer',
+  'X-OTRS-ArticleType' => 'email-internal',
+  'X-OTRS-FollowUp-ArticleType' => 'email-external'
+};
 
 }
+
 1;